Move validation messages into a separate file
